Update string checking logic to use the type information
This simplifies the checking of keys in structured logging by simply looking at the type
information to determine if it's 1) a string 2) a constant.  With this
change, constants in another file and in a different package now work.

Using TypeAndValue fulfills 1 and 2:
1) Check type is string.
2) Check constant. Per https://pkg.go.dev/go/types#TypeAndValue

> TypeAndValue reports the type and value (for constants) of the
corresponding expression.

So if the value of TypeAndValue is present, then it is a constant.